test: moving the test cases to make the partitioning more distinct
