+++ db_test.go
@@ -24,7 +24,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/xujiajun/utils/strconv2"
 )
 
 var (
@@ -200,37 +199,6 @@ func TestDB_BPTSparse(t *testing.T) {
 	})
 }
 
-func TestDB_MergeForString(t *testing.T) {
-	opts := DefaultOptions
-	opts.SegmentSize = 1 * 100
-	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
-		bucket := "bucket"
-		txPut(t, db, bucket, GetTestBytes(0), GetRandomBytes(24), Persistent, nil)
-		txPut(t, db, bucket, GetTestBytes(1), GetRandomBytes(24), Persistent, nil)
-		txDel(t, db, bucket, GetTestBytes(1), nil)
-		txGet(t, db, bucket, GetTestBytes(1), nil, ErrNotFoundKey)
-		validKeyNum := db.BPTreeIdx[bucket].ValidKeyCount
-		require.EqualValuesf(t, 1, validKeyNum, "err GetValidKeyCount. got %d want %d", validKeyNum, 1)
-		require.NoError(t, db.Merge())
-	})
-}
-
-func TestDB_MergeRepeated(t *testing.T) {
-	opts := DefaultOptions
-	opts.SegmentSize = 120
-	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
-		bucket := "bucket"
-		for i := 0; i < 20; i++ {
-			txPut(t, db, bucket, []byte("hello"), []byte("world"), Persistent, nil)
-		}
-		require.Equal(t, int64(9), db.MaxFileID)
-		txGet(t, db, bucket, []byte("hello"), []byte("world"), nil)
-		require.NoError(t, db.Merge())
-		require.Equal(t, int64(10), db.MaxFileID)
-		txGet(t, db, bucket, []byte("hello"), []byte("world"), nil)
-	})
-}
-
 func txSAdd(t *testing.T, db *DB, bucket string, key, value []byte, expectErr error) {
 	err := db.Update(func(tx *Tx) error {
 		err := tx.SAdd(bucket, key, value)
@@ -256,37 +224,6 @@ func txSRem(t *testing.T, db *DB, bucket string, key, value []byte, expectErr er
 		return nil
 	})
 	require.NoError(t, err)
-}
-
-func TestDB_MergeForSet(t *testing.T) {
-	opts := DefaultOptions
-	opts.SegmentSize = 100
-	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
-		bucket := "bucket"
-		key := GetTestBytes(0)
-
-		for i := 0; i < 100; i++ {
-			txSAdd(t, db, bucket, key, GetTestBytes(i), nil)
-		}
-
-		for i := 0; i < 100; i++ {
-			txSIsMember(t, db, bucket, key, GetTestBytes(i), true)
-		}
-
-		for i := 0; i < 50; i++ {
-			txSRem(t, db, bucket, key, GetTestBytes(i), nil)
-		}
-
-		for i := 0; i < 50; i++ {
-			txSIsMember(t, db, bucket, key, GetTestBytes(i), false)
-		}
-
-		for i := 50; i < 100; i++ {
-			txSIsMember(t, db, bucket, key, GetTestBytes(i), true)
-		}
-
-		require.NoError(t, db.Merge())
-	})
 }
 
 func txZAdd(t *testing.T, db *DB, bucket string, key, value []byte, score float64, expectErr error) {
@@ -323,39 +260,6 @@ func txZRangeByRank(t *testing.T, db *DB, bucket string, start, end int) {
 		return nil
 	})
 	require.NoError(t, err)
-}
-
-func TestDB_MergeForZSet(t *testing.T) {
-	opts := DefaultOptions
-	opts.SegmentSize = 100
-	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
-		bucket := "bucket"
-
-		for i := 0; i < 100; i++ {
-			score, _ := strconv2.IntToFloat64(i)
-			txZAdd(t, db, bucket, GetTestBytes(i), GetTestBytes(i), score, nil)
-		}
-
-		for i := 0; i < 100; i++ {
-			txZGetByKey(t, db, bucket, GetTestBytes(i), nil)
-		}
-
-		for i := 0; i < 50; i++ {
-			txZRem(t, db, bucket, GetTestBytes(i), nil)
-		}
-
-		for i := 0; i < 50; i++ {
-			txZGetByKey(t, db, bucket, GetTestBytes(i), ErrNotFoundKey)
-		}
-
-		for i := 50; i < 100; i++ {
-			txZGetByKey(t, db, bucket, GetTestBytes(i), nil)
-		}
-
-		txZRangeByRank(t, db, bucket, 20, 30)
-
-		require.NoError(t, db.Merge())
-	})
 }
 
 func txPop(t *testing.T, db *DB, bucket string, key, expectVal []byte, expectErr error, isLeft bool) {
@@ -405,25 +309,6 @@ func txRange(t *testing.T, db *DB, bucket string, key []byte, start, end, expect
 		return nil
 	})
 	require.NoError(t, err)
-}
-
-func TestDB_MergeForList(t *testing.T) {
-	opts := DefaultOptions
-	opts.SegmentSize = 100
-	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
-		bucket := "bucket"
-		key := GetTestBytes(0)
-
-		for i := 0; i < 100; i++ {
-			txPush(t, db, bucket, key, GetTestBytes(i), nil, true)
-		}
-		txRange(t, db, bucket, key, 0, 99, 100)
-
-		txPop(t, db, bucket, key, GetTestBytes(99), nil, true)
-		txPop(t, db, bucket, key, GetTestBytes(0), nil, false)
-
-		require.NoError(t, db.Merge())
-	})
 }
 
 func TestDB_GetKeyNotFound(t *testing.T) {
