+++ merge_test.go
@@ -16,10 +16,125 @@ package nutsdb
 
 import (
 	"github.com/stretchr/testify/require"
+	"github.com/xujiajun/utils/strconv2"
 	"sync"
 	"testing"
 	"time"
 )
+
+func TestDB_MergeForString(t *testing.T) {
+	opts := DefaultOptions
+	opts.SegmentSize = 1 * 100
+	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
+		bucket := "bucket"
+		txPut(t, db, bucket, GetTestBytes(0), GetRandomBytes(24), Persistent, nil)
+		txPut(t, db, bucket, GetTestBytes(1), GetRandomBytes(24), Persistent, nil)
+		txDel(t, db, bucket, GetTestBytes(1), nil)
+		txGet(t, db, bucket, GetTestBytes(1), nil, ErrNotFoundKey)
+		validKeyNum := db.BPTreeIdx[bucket].ValidKeyCount
+		require.EqualValuesf(t, 1, validKeyNum, "err GetValidKeyCount. got %d want %d", validKeyNum, 1)
+		require.NoError(t, db.Merge())
+	})
+}
+
+func TestDB_MergeRepeated(t *testing.T) {
+	opts := DefaultOptions
+	opts.SegmentSize = 120
+	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
+		bucket := "bucket"
+		for i := 0; i < 20; i++ {
+			txPut(t, db, bucket, []byte("hello"), []byte("world"), Persistent, nil)
+		}
+		require.Equal(t, int64(9), db.MaxFileID)
+		txGet(t, db, bucket, []byte("hello"), []byte("world"), nil)
+		require.NoError(t, db.Merge())
+		require.Equal(t, int64(10), db.MaxFileID)
+		txGet(t, db, bucket, []byte("hello"), []byte("world"), nil)
+	})
+}
+
+func TestDB_MergeForSet(t *testing.T) {
+	opts := DefaultOptions
+	opts.SegmentSize = 100
+	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
+		bucket := "bucket"
+		key := GetTestBytes(0)
+
+		for i := 0; i < 100; i++ {
+			txSAdd(t, db, bucket, key, GetTestBytes(i), nil)
+		}
+
+		for i := 0; i < 100; i++ {
+			txSIsMember(t, db, bucket, key, GetTestBytes(i), true)
+		}
+
+		for i := 0; i < 50; i++ {
+			txSRem(t, db, bucket, key, GetTestBytes(i), nil)
+		}
+
+		for i := 0; i < 50; i++ {
+			txSIsMember(t, db, bucket, key, GetTestBytes(i), false)
+		}
+
+		for i := 50; i < 100; i++ {
+			txSIsMember(t, db, bucket, key, GetTestBytes(i), true)
+		}
+
+		require.NoError(t, db.Merge())
+	})
+}
+
+func TestDB_MergeForZSet(t *testing.T) {
+	opts := DefaultOptions
+	opts.SegmentSize = 100
+	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
+		bucket := "bucket"
+
+		for i := 0; i < 100; i++ {
+			score, _ := strconv2.IntToFloat64(i)
+			txZAdd(t, db, bucket, GetTestBytes(i), GetTestBytes(i), score, nil)
+		}
+
+		for i := 0; i < 100; i++ {
+			txZGetByKey(t, db, bucket, GetTestBytes(i), nil)
+		}
+
+		for i := 0; i < 50; i++ {
+			txZRem(t, db, bucket, GetTestBytes(i), nil)
+		}
+
+		for i := 0; i < 50; i++ {
+			txZGetByKey(t, db, bucket, GetTestBytes(i), ErrNotFoundKey)
+		}
+
+		for i := 50; i < 100; i++ {
+			txZGetByKey(t, db, bucket, GetTestBytes(i), nil)
+		}
+
+		txZRangeByRank(t, db, bucket, 20, 30)
+
+		require.NoError(t, db.Merge())
+	})
+}
+
+func TestDB_MergeForList(t *testing.T) {
+	opts := DefaultOptions
+	opts.SegmentSize = 100
+	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
+		bucket := "bucket"
+		key := GetTestBytes(0)
+
+		for i := 0; i < 100; i++ {
+			txPush(t, db, bucket, key, GetTestBytes(i), nil, true)
+		}
+		txRange(t, db, bucket, key, 0, 99, 100)
+
+		txPop(t, db, bucket, key, GetTestBytes(99), nil, true)
+		txPop(t, db, bucket, key, GetTestBytes(0), nil, false)
+
+		require.NoError(t, db.Merge())
+	})
+}
 
 func TestDB_MergeAutomatic(t *testing.T) {
 	opts := DefaultOptions
@@ -29,8 +144,9 @@ func TestDB_MergeAutomatic(t *testing.T) {
 	bucket := "bucket"
 
 	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
-		_, pendingMergeFileIds := db.getMaxFileIDAndFileIDs()
-		require.Len(t, pendingMergeFileIds, 1)
+		err := db.Merge()
+		require.Error(t, err)
+		require.Error(t, ErrDontNeedMerge, err)
 
 		key := GetTestBytes(0)
 		value := GetRandomBytes(24)
@@ -44,7 +160,7 @@ func TestDB_MergeAutomatic(t *testing.T) {
 		// waiting for the merge work to be triggered.
 		time.Sleep(200 * time.Millisecond)
 
-		_, pendingMergeFileIds = db.getMaxFileIDAndFileIDs()
+		_, pendingMergeFileIds := db.getMaxFileIDAndFileIDs()
 		// because there is only one valid entry, there will be only one data file after merging
 		require.Len(t, pendingMergeFileIds, 1)
 
@@ -114,5 +230,16 @@ func TestDB_MergeWithTx(t *testing.T) {
 				txGet(t, db, bucket, GetTestBytes(i), values[i], nil)
 			}
 		}
+	})
+}
+
+func TestDB_MergeInHintBPTSparseIdxMode(t *testing.T) {
+	opts := DefaultOptions
+	opts.EntryIdxMode = HintBPTSparseIdxMode
+
+	runNutsDBTest(t, &opts, func(t *testing.T, db *DB) {
+		err := db.Merge()
+		require.Error(t, err)
+		require.Equal(t, ErrNotSupportHintBPTSparseIdxMode, err)
 	})
 }
